Add RT-11 fsck/repair

+++ lsi11/tools/rt11tool.c
@@ -123,6 +123,7 @@ static void usage(FILE *out)
             "  add <image> --dir <dir> [--partition N]\n"
             "  rm <image> <NAME.EXT> [--partition N] [--force]\n"
             "  protect <image> <NAME.EXT> [--partition N] [--clear]\n"
+            "  fsck <image> [--partition N] [--repair]\n"
             "  squeeze <image> [--partition N]\n"
             "  mkfs <image> (--blocks N | --rk05 | --rl02 | --rp06 | --rp07)\n"
             "       [--segments N] [--volid TEXT] [--owner TEXT] [--sysid TEXT]\n");
@@ -1105,6 +1106,72 @@ static int cmd_squeeze(int argc, char **argv)
     return 0;
 }
 
+static int cmd_fsck(int argc, char **argv)
+{
+    if (argc < 1) {
+        usage(stderr);
+        return 1;
+    }
+    const char *image = argv[0];
+    uint32_t partition = 0;
+    int have_partition = 0;
+    int repair = 0;
+    int i;
+    rt11_image_t img;
+    unsigned errors = 0;
+    unsigned fixes = 0;
+
+    for (i = 1; i < argc; i++) {
+        if (strcmp(argv[i], "--partition") == 0 && i + 1 < argc) {
+            if (parse_partition_value(argv[i + 1], &partition) != 0) {
+                fprintf(stderr, "rt11tool: invalid partition\n");
+                return 1;
+            }
+            have_partition = 1;
+            i++;
+        } else if (strcmp(argv[i], "--repair") == 0) {
+            repair = 1;
+        } else {
+            usage(stderr);
+            return 1;
+        }
+    }
+
+    if (rt11_open_image(&img, image, repair ? "rb+" : "rb") != 0) {
+        perror("rt11tool: open image");
+        return 1;
+    }
+
+    if (have_partition && rt11_set_partition(&img, partition) != 0) {
+        perror("rt11tool: partition");
+        rt11_close_image(&img);
+        return 1;
+    }
+
+    if (rt11_fsck(&img, repair, stdout, &errors, &fixes) != 0) {
+        perror("rt11tool: fsck");
+        rt11_close_image(&img);
+        return 1;
+    }
+
+    if (errors == 0) {
+        printf("fsck: clean\n");
+    } else if (repair && errors == fixes) {
+        printf("fsck: fixed %u issues\n", fixes);
+    } else {
+        printf("fsck: %u issues found (%u fixed)\n", errors, fixes);
+    }
+
+    rt11_close_image(&img);
+    if (errors == 0) {
+        return 0;
+    }
+    if (repair && errors == fixes) {
+        return 0;
+    }
+    return 1;
+}
+
 static int cmd_mkfs(int argc, char **argv)
 {
     const char *image = argv[0];
@@ -1249,6 +1316,14 @@ int main(int argc, char **argv)
             return 1;
         }
         return cmd_squeeze(argc - 2, &argv[2]);
+    }
+
+    if (strcmp(argv[1], "fsck") == 0) {
+        if (argc < 3) {
+            usage(stderr);
+            return 1;
+        }
+        return cmd_fsck(argc - 2, &argv[2]);
     }
 
     if (strcmp(argv[1], "mkfs") == 0) {
